update: improve code and perf
